+++ HOPSAN++/HopsanCore/Nodes/Nodes.h
@@ -78,7 +78,7 @@ namespace hopsan {
 
         virtual void setSpecialStartValues(Node *pNode)
         {
-            for(int i=0; i<mDataNames.size(); ++i)
+            for(size_t i=0; i<mDataNames.size(); ++i)
             {
                 if(WAVEVARIABLE==i)
                 {
@@ -154,7 +154,7 @@ namespace hopsan {
 
         virtual void setSpecialStartValues(Node *pNode)
         {
-            for(int i=0; i<mDataNames.size(); ++i)
+            for(size_t i=0; i<mDataNames.size(); ++i)
             {
                 if(WAVEVARIABLE==i)
                 {
@@ -189,7 +189,7 @@ namespace hopsan {
 
         virtual void setSpecialStartValues(Node *pNode)
         {
-            for(int i=0; i<mDataNames.size(); ++i)
+            for(size_t i=0; i<mDataNames.size(); ++i)
             {
                 if(WAVEVARIABLE==i)
                 {
@@ -224,7 +224,7 @@ namespace hopsan {
 
         virtual void setSpecialStartValues(Node *pNode)
         {
-            for(int i=0; i<mDataNames.size(); ++i)
+            for(size_t i=0; i<mDataNames.size(); ++i)
             {
                 if(WAVEVARIABLE==i)
                 {
